Create fn read_config for reading toml file
The read_config function is currently considered dead code since it is utilized,
but its values are not employed in the rest of the program. It served as the initial
attempt to read and deserialize the configuration. Additionally, I implemented a
mechanism to handle scenarios where the config file is absent. In such cases,
the read_config function will generate a sample config.toml file.